correção apresentação

+++ lib.c
@@ -107,14 +107,14 @@ node * insertNode(node *current, int info) {
 // obtem item com determinado valor de um conjunto
 node * getNodeByValue(node *first, int value) {
     node *temp, *test = first;
-    do {
+    while (test != NULL) {
         temp = test;
         test = test->next;
 
         if (temp->info == value) {
             return temp;
         }
-    } while (test != NULL);
+    }
 
     return NULL;
 };
@@ -136,7 +136,8 @@ int deleteNode(node *current) {
 // retira itens com determinado valor de um conjunto
 node * deleteNodesWithValue(node *first, int value) {
     node *temp, *test = first;
-    do {
+
+    while (test != NULL) {
         temp = test;
         test = test->next;
 
@@ -152,7 +153,7 @@ node * deleteNodesWithValue(node *first, int value) {
                 first = temp->next;
             free(temp);
         }
-    } while (test != NULL);
+    }
 
     return first;
 };
@@ -161,11 +162,11 @@ node * deleteNodesWithValue(node *first, int value) {
 // retira todos os itens de um conjunto
 int deleteAllNodes(node *first) {
     node *temp, *test = first;
-    do {
+    while (test != NULL) {
         temp = test;
         test = test->next;
         free(temp);
-    } while (test != NULL);
+    }
     return 1;
 };
 
@@ -213,7 +214,8 @@ nodeset * deleteNodeSet(nodeset *current) {
     node *first = current->head;
 
     // libera todos os nodes da memória
-    deleteAllNodes(first);
+    if (first)
+        deleteAllNodes(first);
 
     // reestrutura a lista de conjuntos
     if (current->prev != NULL) {
@@ -231,7 +233,7 @@ nodeset * deleteNodeSet(nodeset *current) {
     }
 
     // se o item excluido for a cabeça, define-se nova cabeça
-    if (test == current && current->next != NULL)
+    if (test == current)
         test = current->next;
 
     // libera da memória o conjunto atual
@@ -373,37 +375,37 @@ node * executeOp(nodeset *head, int op) {
 // recebe duas cabeças de listas e retorna a cabeça de uma terceira lista que é a união das duas entradas
 node * getUnion(node *head1, node *head2) {
     // node1 e node2 recebem as cabeças para serem tratadas na função
-	node *node1 = head1;
-	node *node2 = head2;
-	// head3 será a cabeça da nova lista gerada
-	// lastNode será o último node adicionado, para que o próximo seja colocado depois dele
-	// prev1 e prev2 receberão os nodes visitados anteriormente
-	// a fim de que não haja itens repetidos no resultado
-	node *head3, *lastNode, *prev1, *prev2;
+    node *node1 = head1;
+    node *node2 = head2;
+    // head3 será a cabeça da nova lista gerada
+    // lastNode será o último node adicionado, para que o próximo seja colocado depois dele
+    // prev1 e prev2 receberão os nodes visitados anteriormente
+    // a fim de que não haja itens repetidos no resultado
+    node *head3, *lastNode, *prev1, *prev2;
 
-	head3 = lastNode = prev1 = prev2 = initNode();
+    head3 = lastNode = prev1 = prev2 = initNode();
 
     // caso haja itens em ambas as listas dadas como entrada
     // percorre as duas listas até o fim da menor delas
-	while (node1 && node2) {
+    while (node1 && node2) {
         // se os valores de node1 e node2 forem iguais
-		if (node1->info == node2->info) {
+        if (node1->info == node2->info) {
             // se for o primeiro item, adiciona-se e define-se a cabeça da lista resultado
-			if (prev1 == NULL) {
+            if (prev1 == NULL) {
                 lastNode = insertNode(lastNode, node1->info);
-				if (head3 == NULL)
+                if (head3 == NULL)
                     head3 = lastNode;
             // se não, adiciona apenas se for diferente do node visitado anteriormente
-			} else {
-				if(prev1->info != node1->info)
+            } else {
+                if(prev1->info != node1->info)
                     lastNode = insertNode(lastNode, node1->info);
-			}
-			// atualiza itens visitados anteriormente e próximos
-			prev1 = node1;
-			prev2 = node2;
-			node1 = node1->next;
-			node2 = node2->next;
-		} else {
+            }
+            // atualiza itens visitados anteriormente e próximos
+            prev1 = node1;
+            prev2 = node2;
+            node1 = node1->next;
+            node2 = node2->next;
+        } else {
             // quando forem diferentes, o procedimento é o mesmo, a única diferença é que
             // como as listas estão organizadas com mergeSort, avançamos somente a lista mais "atrasada", ou seja, que tem o menor valor
             if (node1->info < node2->info) {
@@ -431,32 +433,32 @@ node * getUnion(node *head1, node *head2) {
                 prev2 = node2;
                 node2 = node2->next;
             }
-		}
-	}
+        }
+    }
 
     // caso uma das listas de entrada tenha acabado de ser percorrida
     // então o resultado receberá o restante da outra lista, mas sem itens repetidos
-	if (node1 == NULL) {
+    if (node1 == NULL) {
         // caso node1 seja NULL vamos percorrer node2
-		while (node2) {
+        while (node2) {
             // se prev2 for NULL significa que estamos no primeiro elemento de node2
             // então basta adicionar o elemento
-			if (prev2 == NULL) {
-				lastNode = insertNode(lastNode, node2->info);
-				// CASO UMA DAS LISTAS SEJA NULL NÃO HAVERÁ CABEÇA DO RESULTADO DEFINIDA
-				if (head3 == NULL)
+            if (prev2 == NULL) {
+                lastNode = insertNode(lastNode, node2->info);
+                // CASO UMA DAS LISTAS SEJA NULL NÃO HAVERÁ CABEÇA DO RESULTADO DEFINIDA
+                if (head3 == NULL)
                     head3 = lastNode;
-			} else {
+            } else {
                 // a partir do segundo elemento, verificamos se ele é diferente do que
                 // foi visitado anteriormente para ele não seja repetido
-				if (prev2->info != node2->info)
+                if (prev2->info != node2->info)
                     lastNode = insertNode(lastNode, node2->info);
-			}
-			// atualizamos o anterior como o que acabamos de visitar e o próximo a ser visitado
-			prev2 = node2;
-			node2 = node2->next;
-		}
-	} else {
+            }
+            // atualizamos o anterior como o que acabamos de visitar e o próximo a ser visitado
+            prev2 = node2;
+            node2 = node2->next;
+        }
+    } else {
         // igual o caso anterior, só que usando node1
         if (node2 == NULL) {
             while (node1) {
@@ -472,9 +474,9 @@ node * getUnion(node *head1, node *head2) {
                 node1 = node1->next;
             }
         }
-	}
+    }
 
-	return head3;
+    return head3;
 }
 
 
@@ -482,30 +484,30 @@ node * getUnion(node *head1, node *head2) {
 // recebe duas cabeças de listas e retorna uma nova lista
 node * getInt(node *head1, node *head2) {
     // node1 e node2 recebem as cabeças para serem tratadas na função
-	node *node1 = head1;
-	node *node2 = head2;
-	// head3 será a cabeça da nova lista gerada
-	// lastNode será o último node adicionado, para colocar um depois do outro
-	node *head3, *lastNode;
+    node *node1 = head1;
+    node *node2 = head2;
+    // head3 será a cabeça da nova lista gerada
+    // lastNode será o último node adicionado, para colocar um depois do outro
+    node *head3, *lastNode;
 
-	head3 = lastNode = initNode();
+    head3 = lastNode = initNode();
 
     // caso uma das cabeças seja NULL não há como realizar intersecção
-	if(head1 == NULL || head2 == NULL)
-		return NULL;
+    if(head1 == NULL || head2 == NULL)
+        return NULL;
 
     // percorremos as listas até o fim da menor delas
-	while(node1 && node2) {
+    while(node1 && node2) {
         // se os dados forem iguais, adicionamos à intersecção
-		if(node1->info == node2->info) {
+        if(node1->info == node2->info) {
             lastNode = insertNode(lastNode, node1->info);
             // se a lista resultado estiver vazia, ela recebe o primeiro node adicionado
             if (head3 == NULL)
                 head3 = lastNode;
-			node1 = node1->next;
-			node2 = node2->next;
+            node1 = node1->next;
+            node2 = node2->next;
 
-		} else {
+        } else {
             // como as listas estão organizadas com mergeSort
             // quando os valores forem diferentes, avançamos a lista mais "atrasada", ou seja, que tem o menor valor
             if(node1->info < node2->info)
@@ -513,9 +515,9 @@ node * getInt(node *head1, node *head2) {
             else
                 node2 = node2->next;
         }
-	}
+    }
 
-	return head3;
+    return head3;
 }
 
 
@@ -523,11 +525,11 @@ node * getInt(node *head1, node *head2) {
 // e retorna uma nova lista
 // basicamente subtrai a intersecção da união
 node * getDif(nodeset *head) {
-	node *union_list, *intersection_list, *diff_list, *test;
+    node *union_list, *intersection_list, *diff_list, *test;
 
     // gera união e intersecção
-	union_list = executeOp(head, 1);
-	intersection_list = executeOp(head, 2);
+    union_list = executeOp(head, 1);
+    intersection_list = executeOp(head, 2);
 
     // o resultado recebe a união e uma variável de testes é instanciada para percorrer
     // a lista de intersecção
@@ -538,13 +540,13 @@ node * getDif(nodeset *head) {
     while (test != NULL) {
         diff_list = deleteNodesWithValue(diff_list, test->info);
         test = test->next;
-	}
+    }
 
     // limpamos da memória a lista de intersecção, pois não é mais necessária
     if (intersection_list != NULL)
         deleteAllNodes(intersection_list);
 
-	return diff_list;
+    return diff_list;
 }
